paddlemix/requirements.txt
numpy
tensorboardX
opencv-python
Pillow
ftfy
regex
einops>=0.6.1
tiktoken
